requirements.txt
torch~=2.1.2
matplotlib~=3.8.2
pillow~=10.2.0
scikit-learn~=1.4.0
torchmetrics~=1.3.0.post0
torchvision~=0.16.2
numpy~=1.26.3